+++ src/main/java/util/Time.java
@@ -2,7 +2,7 @@ package util;
 
 public class Time {
     public static float timeStarted = System.nanoTime();
-    public static float getTime() {
-        return (float)((System.nanoTime() - timeStarted) * 1E-9);
+    public static double getTime() {
+        return ((System.nanoTime() - timeStarted) * 1E-9);
     }
 }
